+++ react-components/src/actions/ActionTypes.js
@@ -7,5 +7,6 @@ export const SET_CHANNEL_SYNC_MODAL_OPEN = 'SET_CHANNEL_SYNC_MODAL_OPEN';
 export const SET_CHANNELS = 'SET_CHANNELS';
 export const SET_DB_STALE = 'SET_DB_STALE';
 
+/* ---- CHANNEL MANAGER ---- */
 export const SET_ACTIVE_CHANNEL_VIDEOS = 'SET_ACTIVE_CHANNEL_VIDEOS';
 export const SET_ACTIVE_CHANNEL_VIDEOS_LOADING = 'SET_ACTIVE_CHANNEL_VIDEOS_LOADING';
