+++ react-components/src/actions/Globals.js
@@ -1,0 +1,5 @@
+import * as aTypes from './ActionTypes';
+
+export const setSessionAction = (data) => async (dispatch) => {
+  dispatch({ type: aTypes.SET_SESSION, payload: data });
+};
